Refactor AIR detection: fix timestamp logic, optimize polling, add hysteresis mechanism

+++ src/config.h
@@ -28,7 +28,7 @@ typedef struct {
     uint8_t tof_offset;  // 起始高度, 默认 60mm
     uint8_t tof_pitch;   // 每段高度 (Air1-5), 默认 30mm
     uint8_t air6_range;  // Air6 检测范围, 默认 150mm (>= pitch)
-    uint8_t air_min_hold_ms;  // AIR 最小按下持续时间, 默认 50ms
+    uint16_t air_min_hold_ms;  // AIR 最小按下持续时间, 默认 50ms (修正为 uint16_t)
 
     // 保留原有阈值配置 (兼容性)
     uint16_t air_threshold[5];
